requirements.txt
APScheduler==3.7.0
netifaces==0.10.9
pycparser==2.20
PyYAML==5.4.1
sanic==20.12.2
SQLAlchemy==1.3.23
python3-nmap
python-libnmap
redis